% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{bijector_chain}
\alias{bijector_chain}
\title{Bijector which applies a sequence of bijectors.}
\usage{
bijector_chain(bijectors = NULL, validate_args = FALSE, name = NULL)
}
\arguments{
\item{bijectors}{`list` of bijector instances. An empty list makes this
bijector equivalent to the `Identity` bijector.}

\item{validate_args}{Logical indicating whether arguments should be checked for correctness.}

\item{name}{String, name given to ops managed by this object. Default:
E.g., `Chain([Exp(), Softplus()]).name == "chain_of_exp_of_softplus"`.}
}
\description{
Bijector which applies a sequence of bijectors.
}
