% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{bijector_cholesky_to_inv_cholesky}
\alias{bijector_cholesky_to_inv_cholesky}
\title{Title}
\usage{
bijector_cholesky_to_inv_cholesky(validate_args = FALSE,
  name = "cholesky_to_inv_cholesky")
}
\arguments{
\item{name}{}
}
\description{
Title
}
