% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{bijector_sigmoid}
\alias{bijector_sigmoid}
\title{Bijector which computes `Y = g(X) = 1 / (1 + exp(-X))`.}
\usage{
bijector_sigmoid(validate_args = FALSE, name = "sigmoid")
}
\arguments{
\item{validate_args}{Logical, default `FALSE`. Whether to validate input with asserts. If `validate_args` is
`FALSE`, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
Bijector which computes `Y = g(X) = 1 / (1 + exp(-X))`.
}
