% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{cdf.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{cdf.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Cumulative distribution function.
Given random variable `X`, the cumulative distribution function `cdf` is:
`cdf(x) := P[X <= x]``}
\usage{

  \method{cdf}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  value, name = "cdf")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{value}{`float` or `double` `Tensor`.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Cumulative distribution function.
Given random variable `X`, the cumulative distribution function `cdf` is:
`cdf(x) := P[X <= x]``
}
