% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distribution_transformed}
\alias{distribution_transformed}
\title{A Transformed Distribution.}
\usage{
distribution_transformed(distribution, bijector, batch_shape = NULL,
  event_shape = NULL, validate_args = FALSE, name = NULL)
}
\arguments{
\item{distribution}{The base distribution instance to transform. Typically an instance of `Distribution`.}

\item{bijector}{The object responsible for calculating the transformation. Typically an instance of `Bijector`.}

\item{batch_shape}{`integer` vector `Tensor` which overrides `distribution` `batch_shape`;
valid only if `distribution.is_scalar_batch()`.}

\item{event_shape}{`integer` vector `Tensor` which overrides `distribution` `event_shape`;
valid only if `distribution.is_scalar_event()`.}

\item{validate_args}{Logical, default `FALSE`. When `TRUE` distribution parameters are checked
for validity despite possibly degrading runtime performance. When `FALSE` invalid inputs may
silently render incorrect outputs. Default value: `FALSE`.}

\item{name}{The name for ops managed by the distribution.  Default value: `bijector.name + distribution.name`.}
}
\description{
A `TransformedDistribution` models `p(y)` given a base distribution `p(x)`,
and a deterministic, invertible, differentiable transform, `Y = g(X)`. The
transform is typically an instance of the `Bijector` class and the base
distribution is typically an instance of the `Distribution` class.
}
