% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{entropy.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{entropy.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Shannon entropy in nats.}
\usage{

  \method{entropy}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  name = "entropy")
}
\description{
Shannon entropy in nats.
}
