% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijector-methods.R
\name{forward.tensorflow_probability.python.bijectors.bijector.Bijector}
\alias{forward.tensorflow_probability.python.bijectors.bijector.Bijector}
\title{Returns the forward Bijector evaluation, i.e., X = g(Y).}
\usage{

  \method{forward}{tensorflow_probability.python.bijectors.bijector.Bijector}(bijector,
  x, name = "forward")
}
\arguments{
\item{bijector}{The bijector to apply}

\item{x}{Tensor. The input to the "forward" evaluation.}

\item{name}{The name to give this op.}
}
\value{
Tensor
}
\description{
Returns the forward Bijector evaluation, i.e., X = g(Y).
}
