% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijector-methods.R
\name{forward_log_det_jacobian.tensorflow_probability.python.bijectors.bijector.Bijector}
\alias{forward_log_det_jacobian.tensorflow_probability.python.bijectors.bijector.Bijector}
\title{Returns both the forward_log_det_jacobian.}
\usage{

  \method{forward_log_det_jacobian}{tensorflow_probability.python.bijectors.bijector.Bijector}(bijector,
  x, event_ndims, name = "forward_log_det_jacobian")
}
\arguments{
\item{bijector}{The bijector to apply}

\item{x}{Tensor. The input to the "forward" Jacobian determinant evaluation.}

\item{event_ndims}{Number of dimensions in the probabilistic events being transformed.
Must be greater than or equal to bijector$forward_min_event_ndims. The result is summed over the final
dimensions to produce a scalar Jacobian determinant for each event, i.e. it has shape
 x$shape$ndims - event_ndims dimensions.}

\item{name}{The name to give this op.}
}
\value{
Tensor, if this bijector is injective. If not injective this is not implemented.
}
\description{
Returns both the forward_log_det_jacobian.
}
