% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_fit_one_step}
\alias{glm_fit_one_step}
\title{Runs one Fisher scoring step}
\usage{
glm_fit_one_step(x, ...)
}
\arguments{
\item{x}{float-like, matrix-shaped Tensor where each row represents a sample's
features.}

\item{...}{other arguments passed to specific methods.}
}
\value{
A \code{glm_fit} object with parameter estimates, number of iterations,
etc.
}
\description{
Runs one Fisher scoring step
}
\seealso{
\code{\link[=glm_fit_one_step.tensorflow.tensor]{glm_fit_one_step.tensorflow.tensor()}}
}
