% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{initializer_blockwise}
\alias{initializer_blockwise}
\title{Blockwise Initializer}
\usage{
initializer_blockwise(initializers, sizes, validate_args = FALSE)
}
\arguments{
\item{initializers}{list of Keras initializers, eg: \code{\link[keras:initializer_glorot_uniform]{keras::initializer_glorot_uniform()}}
or \code{\link[=initializer_constant]{initializer_constant()}}.}

\item{sizes}{list of integers scalars representing the number of elements associated
with each initializer in \code{initializers}.}

\item{validate_args}{bool indicating we should do (possibly expensive) graph-time
assertions, if necessary.

@return Initializer which concats other intializers}
}
\description{
Initializer which concats other intializers
}
