% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-layers.R
\name{layer_categorical_mixture_of_one_hot_categorical}
\alias{layer_categorical_mixture_of_one_hot_categorical}
\title{A OneHotCategorical mixture Keras layer from \code{k * (1 + d)} params.}
\usage{
layer_categorical_mixture_of_one_hot_categorical(object, event_size,
  num_components,
  convert_to_tensor_fn = tfp$distributions$Distribution$sample,
  sample_dtype = NULL, validate_args = FALSE, ...)
}
\arguments{
\item{object}{Model or layer object}

\item{event_size}{Scalar \code{integer} representing the size of single draw from this distribution.}

\item{num_components}{Scalar \code{integer} representing the number of mixture
components. Must be at least 1. (If \code{num_components=1}, it's more
efficient to use the \code{OneHotCategorical} layer.)}

\item{convert_to_tensor_fn}{A callable that takes a tfd$Distribution instance and returns a
tf$Tensor-like object. Default value: \code{tfd$distributions$Distribution$sample}.}

\item{sample_dtype}{\code{dtype} of samples produced by this distribution.
Default value: \code{NULL} (i.e., previous layer's \code{dtype}).}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{...}{Additional arguments passed to \code{args} of \code{keras::create_layer}.}
}
\value{
a Keras layer
}
\description{
\code{k} (i.e., \code{num_components}) represents the number of component
\code{OneHotCategorical} distributions and \code{d} (i.e., \code{event_size}) represents the
number of categories within each \code{OneHotCategorical} distribution.
}
\details{
Typical choices for \code{convert_to_tensor_fn} include:
\itemize{
\item \code{tfp$distributions$Distribution$sample}
\item \code{tfp$distributions$Distribution$mean}
\item \code{tfp$distributions$Distribution$mode}
}
}
\seealso{
For an example how to use in a Keras model, see \code{\link[=layer_independent_normal]{layer_independent_normal()}}.

Other distribution_layers: \code{\link{layer_distribution_lambda}},
  \code{\link{layer_independent_bernoulli}},
  \code{\link{layer_independent_logistic}},
  \code{\link{layer_independent_normal}},
  \code{\link{layer_independent_poisson}},
  \code{\link{layer_kl_divergence_add_loss}},
  \code{\link{layer_kl_divergence_regularizer}},
  \code{\link{layer_mixture_logistic}},
  \code{\link{layer_mixture_normal}},
  \code{\link{layer_mixture_same_family}},
  \code{\link{layer_multivariate_normal_tri_l}},
  \code{\link{layer_one_hot_categorical}}
}
\concept{distribution_layers}
