% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-layers.R
\name{layer_multivariate_normal_tri_l}
\alias{layer_multivariate_normal_tri_l}
\title{A d-variate Multivariate Normal TriL Keras layer from \code{d+d*(d+1)/ 2} params}
\usage{
layer_multivariate_normal_tri_l(object, event_size,
  convert_to_tensor_fn = tfp$distributions$Distribution$sample,
  validate_args = FALSE, batch_input_shape = NULL,
  input_shape = NULL, batch_size = NULL, dtype = NULL, name = NULL,
  trainable = NULL, weights = NULL)
}
\arguments{
\item{object}{Model or layer object}

\item{event_size}{Integer vector tensor representing the shape of single draw from this distribution.}

\item{convert_to_tensor_fn}{A callable that takes a tfd$Distribution instance and returns a
tf$Tensor-like object. Default value: \code{tfd$distributions$Distribution$sample}.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{batch_input_shape}{Shapes, including the batch size. For instance,
\code{batch_input_shape=c(10, 32)} indicates that the expected input will be
batches of 10 32-dimensional vectors. \code{batch_input_shape=list(NULL, 32)}
indicates batches of an arbitrary number of 32-dimensional vectors.}

\item{input_shape}{Dimensionality of the input (integer) not including the
samples axis. This argument is required when using this layer as the first
layer in a model.}

\item{batch_size}{Fixed batch size for layer}

\item{dtype}{The data type expected by the input, as a string (\code{float32},
\code{float64}, \code{int32}...)}

\item{name}{An optional name string for the layer. Should be unique in a
model (do not reuse the same name twice). It will be autogenerated if it
isn't provided.}

\item{trainable}{Whether the layer weights will be updated during training.}

\item{weights}{Initial weights for layer.}
}
\value{
a Keras layer that wraps a MultivariateNormalTriL distribution
}
\description{
A d-variate Multivariate Normal TriL Keras layer from \code{d+d*(d+1)/ 2} params
}
\seealso{
Other distribution_layers: \code{\link{layer_categorical_mixture_of_one_hot_categorical}},
  \code{\link{layer_distribution_lambda}},
  \code{\link{layer_independent_bernoulli}},
  \code{\link{layer_kl_divergence_add_loss}},
  \code{\link{layer_kl_divergence_regularizer}},
  \code{\link{layer_one_hot_categorical}}
}
\concept{distribution_layers}
