% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-layers.R
\name{layer_multivariate_normal_tril}
\alias{layer_multivariate_normal_tril}
\title{A `d`-variate MVNTriL Keras layer from `d+d*(d+1)/ 2` params.}
\usage{
layer_multivariate_normal_tril(object, event_size,
  convert_to_tensor_fn = tfp$distributions$Distribution$sample,
  validate_args = FALSE, batch_input_shape = NULL,
  input_shape = NULL, batch_size = NULL, dtype = NULL, name = NULL,
  trainable = NULL, weights = NULL)
}
\arguments{
\item{event_size}{Integer vector tensor representing the shape of single draw from this distribution.}

\item{convert_to_tensor_fn}{A callable that takes a `tfd$Distribution` instance and returns a
`tf$Tensor`-like object. Default value: `tfd$Distribution$sample`.}

\item{validate_args}{Logical, default `FALSE`. When `TRUE` distribution parameters are checked
for validity despite possibly degrading runtime performance. When `FALSE` invalid inputs may
silently render incorrect outputs. Default value: `FALSE`.}
}
\value{
a Keras layer that wraps a MultivariateNormalTriL distribution
}
\description{
A `d`-variate MVNTriL Keras layer from `d+d*(d+1)/ 2` params.
}
\seealso{
Other Probabilistic layers (require TensorFlow probability): \code{\link{layer_independent_bernoulli}},
  \code{\link{layer_kl_divergence_add_loss}}
}
\concept{Probabilistic layers (require TensorFlow probability)}
