% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{log_cdf.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{log_cdf.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Log cumulative distribution function.}
\usage{

  \method{log_cdf}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  value, name = "log_cdf")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{value}{`float` or `double` `Tensor`.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Given random variable `X`, the cumulative distribution function `cdf` is:
`log_cdf(x) := Log[ P[X <= x] ]``
Often, a numerical approximation can be used for `log_cdf(x)` that yields
a more accurate answer than simply taking the logarithm of the `cdf` when `x << -1`.
}
