% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{log_survival_function.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{log_survival_function.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Log survival function.}
\usage{

  \method{log_survival_function}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  value, name = "log_survival_function")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{value}{`float` or `double` `Tensor`.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Given random variable `X`, the survival function is defined:
`log_survival_function(x) = Log[ P[X > x] ] = Log[ 1 - P[X <= x] ] = Log[ 1 - cdf(x) ]`
}
\details{
Typically, different numerical approximations can be used for the log survival function,
 which are more accurate than `1 - cdf(x)` when `x >> 1`.
}
