% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{mean.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{mean.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Mean.}
\usage{

  \method{mean}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  name = "mean")
}
\arguments{
\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Mean.
}
