% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-static-methods.R
\name{params_size_categorical_mixture_of_one_hot_categorical}
\alias{params_size_categorical_mixture_of_one_hot_categorical}
\title{number of \code{params} needed to create a CategoricalMixtureOfOneHotCategorical distribution}
\usage{
params_size_categorical_mixture_of_one_hot_categorical(
  event_size,
  num_components
)
}
\arguments{
\item{event_size}{event size of this distribution}

\item{num_components}{number of components in the mixture}
}
\value{
a scalar
}
\description{
number of \code{params} needed to create a CategoricalMixtureOfOneHotCategorical distribution
}
