% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-static-methods.R
\name{params_size_independent_logistic}
\alias{params_size_independent_logistic}
\title{number of \code{params} needed to create an IndependentLogistic distribution}
\usage{
params_size_independent_logistic(event_size)
}
\arguments{
\item{event_size}{event size of this distribution}
}
\value{
a scalar
}
\description{
number of \code{params} needed to create an IndependentLogistic distribution
}
