% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-static-methods.R
\name{params_size_mixture_normal}
\alias{params_size_mixture_normal}
\title{number of \code{params} needed to create a MixtureNormal distribution}
\usage{
params_size_mixture_normal(num_components, event_shape)
}
\arguments{
\item{num_components}{Number of component distributions in the mixture distribution.}

\item{event_shape}{Number of parameters needed to create a single component distribution.}
}
\value{
a scalar
}
\description{
number of \code{params} needed to create a MixtureNormal distribution
}
