% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-static-methods.R
\name{params_size_mixture_same_family}
\alias{params_size_mixture_same_family}
\title{number of \code{params} needed to create a MixtureSameFamily distribution}
\usage{
params_size_mixture_same_family(num_components, component_params_size)
}
\arguments{
\item{num_components}{Number of component distributions in the mixture distribution.}

\item{component_params_size}{Number of parameters needed to create a single component distribution.}
}
\value{
a scalar
}
\description{
number of \code{params} needed to create a MixtureSameFamily distribution
}
