% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-static-methods.R
\name{params_size_multivariate_normal_tri_l}
\alias{params_size_multivariate_normal_tri_l}
\title{number of \code{params} needed to create a MultivariateNormalTriL distribution}
\usage{
params_size_multivariate_normal_tri_l(event_size)
}
\arguments{
\item{event_size}{event size of this distribution}
}
\value{
a scalar
}
\description{
number of \code{params} needed to create a MultivariateNormalTriL distribution
}
