% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-static-methods.R
\name{params_size_one_hot_categorical}
\alias{params_size_one_hot_categorical}
\title{number of \code{params} needed to create a OneHotCategorical distribution}
\usage{
params_size_one_hot_categorical(event_size)
}
\arguments{
\item{event_size}{event size of this distribution}
}
\value{
a scalar
}
\description{
number of \code{params} needed to create a OneHotCategorical distribution
}
