% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{prob.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{prob.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Probability density/mass function.}
\usage{

  \method{prob}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  value, name = "prob")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{value}{`float` or `double` `Tensor`.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Probability density/mass function.
}
