% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tf}
\alias{shape}
\alias{tf_config}
\alias{\%>\%}
\title{Objects exported from other packages}
\value{
a alias for tensorflow::tf

a alias for tensorflow::shape

a alias for tensorflow::tf_config

a alias for magrittr::\verb{\%>\%}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tensorflow}{\code{\link[tensorflow]{shape}}, \code{\link[tensorflow]{tf}}, \code{\link[tensorflow]{tf_config}}}
}}

