% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{survival_function.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{survival_function.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Survival function.}
\usage{

  \method{survival_function}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  value, name = "survival_function")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{value}{`float` or `double` `Tensor`.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Given random variable `X`, the survival function is defined:
`survival_function(x) = P[X > x] = 1 - P[X <= x] = 1 - cdf(x).``
}
