% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_affine_linear_operator}
\alias{tfb_affine_linear_operator}
\title{Compute Y = g(X; shift, scale) = scale @ X + shift.}
\usage{
tfb_affine_linear_operator(shift = NULL, scale = NULL,
  adjoint = FALSE, validate_args = FALSE,
  name = "affine_linear_operator")
}
\arguments{
\item{shift}{Floating-point Tensor.}

\item{scale}{Subclass of LinearOperator. Represents the (batch) positive definite matrix M in R^{k x k}.}

\item{adjoint}{Logical indicating whether to use the scale matrix as specified or its adjoint.
Default value: False.}

\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
shift is a numeric Tensor and scale is a LinearOperator.
If X is a scalar then the forward transformation is: scale * X + shift
where * denotes broadcasted elementwise product.
}
