% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_cholesky_to_inv_cholesky}
\alias{tfb_cholesky_to_inv_cholesky}
\title{Maps the Cholesky factor of `M` to the Cholesky factor of `M^{-1}`.}
\usage{
tfb_cholesky_to_inv_cholesky(validate_args = FALSE,
  name = "cholesky_to_inv_cholesky")
}
\arguments{
\item{validate_args}{Logical, default `FALSE`. Whether to validate input with asserts. If `validate_args` is
`FALSE`, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The `forward` and `inverse` calculations are conceptually identical to:
`def forward(x): return tf.cholesky(tf.linalg.inv(tf.matmul(x, x, adjoint_b=True)))``
`inverse = forward`
However, the actual calculations exploit the triangular structure of the matrices.
}
