% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_cumsum}
\alias{tfb_cumsum}
\title{Computes the cumulative sum of a tensor along a specified axis.}
\usage{
tfb_cumsum(axis = -1, validate_args = FALSE, name = "cumsum")
}
\arguments{
\item{axis}{\code{int} indicating the axis along which to compute the cumulative sum.
Note that positive (and zero) values are not supported}
}
\description{
Computes the cumulative sum of a tensor along a specified axis.
}
