% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijector-methods.R
\name{tfb_forward}
\alias{tfb_forward}
\title{Returns the forward Bijector evaluation, i.e., X = g(Y).}
\usage{
tfb_forward(bijector, x, name = "forward")
}
\arguments{
\item{bijector}{The bijector to apply}

\item{x}{Tensor. The input to the "forward" evaluation.}
}
\description{
Returns the forward Bijector evaluation, i.e., X = g(Y).
}
