% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_identity}
\alias{tfb_identity}
\title{Compute Y = g(X) = X.}
\usage{
tfb_identity(validate_args = FALSE, name = "identity")
}
\arguments{
\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
Compute Y = g(X) = X.
}
\concept{bijectors}
