% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijector-methods.R
\name{tfb_inverse}
\alias{tfb_inverse}
\title{Returns the inverse Bijector evaluation, i.e., X = g^{-1}(Y).}
\usage{
tfb_inverse(bijector, y, name = "inverse")
}
\arguments{
\item{bijector}{The bijector to apply}

\item{y}{Tensor. The input to the "inverse" evaluation.}
}
\description{
Returns the inverse Bijector evaluation, i.e., X = g^{-1}(Y).
}
