% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_iterated_sigmoid_centered}
\alias{tfb_iterated_sigmoid_centered}
\title{Bijector which applies a Stick Breaking procedure.}
\usage{
tfb_iterated_sigmoid_centered(validate_args = FALSE,
  name = "iterated_sigmoid")
}
\arguments{
\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
a bijector instance.
}
\description{
Bijector which applies a Stick Breaking procedure.
}
