% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_matrix_inverse_tri_l}
\alias{tfb_matrix_inverse_tri_l}
\title{Computes g(L) = inv(L), where L is a lower-triangular matrix.}
\usage{
tfb_matrix_inverse_tri_l(validate_args = FALSE,
  name = "matrix_inverse_tril")
}
\arguments{
\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
L must be nonsingular; equivalently, all diagonal entries of L must be nonzero.
The input must have rank >= 2.  The input is treated as a batch of matrices
with batch shape input.shape[:-2], where each matrix has dimensions
input.shape[-2] by input.shape[-1] (hence input.shape[-2] must equal input.shape[-1]).
}
