% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_matvec_lu}
\alias{tfb_matvec_lu}
\title{Matrix-vector multiply using LU decomposition.}
\usage{
tfb_matvec_lu(lower_upper, permutation, validate_args = FALSE,
  name = NULL)
}
\arguments{
\item{lower_upper}{The LU factorization as returned by `tf$linalg$lu`.}

\item{permutation}{The LU factorization permutation as returned by `tf$linalg$lu`.}

\item{validate_args}{Logical, default `FALSE`. Whether to validate input with asserts. If `validate_args` is
`FALSE`, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
This bijector is identical to the "Convolution1x1" used in Glow [(Kingma and Dhariwal, 2018)[1].
}
\details{
Warning: this bijector never verifies the scale matrix (as parameterized by LU
ecomposition) is invertible. Ensuring this is the case is the caller's responsibility.

References
[1]: Diederik P. Kingma, Prafulla Dhariwal. Glow: Generative Flow with Invertible 1x1 Convolutions.
_arXiv preprint arXiv:1807.03039_, 2018. https://arxiv.org/abs/1807.03039
}
