% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_ordered}
\alias{tfb_ordered}
\title{Bijector which maps a tensor x_k that has increasing elements in the last dimension to an unconstrained tensor y_k.}
\usage{
tfb_ordered(validate_args = FALSE, name = "ordered")
}
\arguments{
\item{validate_args}{Logical, default `FALSE`. Whether to validate input with asserts. If `validate_args` is
`FALSE`, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
Both the domain and the codomain of the mapping is [-inf, inf], however,
the input of the forward mapping must be strictly increasing.
The inverse of the bijector applied to a normal random vector `y ~ N(0, 1)`
gives back a sorted random vector with the same distribution `x ~ N(0, 1)`
where `x = sort(y)`
}
\details{
On the last dimension of the tensor, Ordered bijector performs:
```y[0] = x[0]```
```y[1:] = tf$log(x[1:] - x[:-1])```'
}
