% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_pad}
\alias{tfb_pad}
\title{Pads a value to the \code{event_shape} of a \code{Tensor}.}
\usage{
tfb_pad(paddings = list(c(0, 1)), mode = "CONSTANT",
  constant_values = 0, axis = NULL, validate_args = FALSE,
  name = NULL)
}
\arguments{
\item{paddings}{A vector-shaped \code{Tensor} of \code{integer} pairs representing the number
of elements to pad on the left and right, respectively.
Default value: \code{list(reticulate::tuple(0L, 1L))}.}

\item{mode}{One of \code{'CONSTANT'}, \code{'REFLECT'}, or \code{'SYMMETRIC'}
(case-insensitive). For more details, see \code{tf$pad}.}

\item{constant_values}{In "CONSTANT" mode, the scalar pad value to use. Must be
same type as \code{tensor}. For more details, see \code{tf$pad}.}

\item{axis}{The dimensions for which \code{paddings} are applied. Must be 1:1 with
\code{paddings} or \code{NULL}.
Default value: \code{NULL} (i.e., \code{tf$range(start = -length(paddings), limit = 0)}).}

\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
a bijector instance.
}
\description{
The semantics of \code{bijector_pad} generally follow that of \code{tf$pad()}
except that \code{bijector_pad}'s \code{paddings} argument applies to the rightmost
dimensions. Additionally, the new argument \code{axis} enables overriding the
dimensions to which \code{paddings} is applied. Like \code{paddings}, the \code{axis}
argument is also relative to the rightmost dimension and must therefore be
negative.
The argument \code{paddings} is a vector of \code{integer} pairs each representing the
number of left and/or right \code{constant_values} to pad to the corresponding
righmost dimensions. That is, unless \code{axis} is specified\code{, specifiying }k\code{different}paddings\code{means the rightmost}k\code{dimensions will be "grown" by the sum of the respective}paddings\code{row. When}axis\code{is specified, it indicates the dimension to which the corresponding}paddings\code{element is applied. By default}axis\code{is}NULL\code{which means it is logically equivalent to}range(start=-len(paddings), limit=0)`, i.e., the rightmost dimensions.
}
