% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_reshape}
\alias{tfb_reshape}
\title{Reshapes the event_shape of a Tensor.}
\usage{
tfb_reshape(event_shape_out, event_shape_in = c(-1),
  validate_args = FALSE, name = NULL)
}
\arguments{
\item{event_shape_out}{An integer-like vector-shaped Tensor
representing the event shape of the transformed output.}

\item{event_shape_in}{An optional integer-like vector-shape Tensor
representing the event shape of the input. This is required in
order to define inverse operations; the default of list(-1) assumes a vector-shaped input.}

\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The semantics generally follow that of tf.reshape(), with a few differences:
  * The user must provide both the input and output shape, so that
    the transformation can be inverted. If an input shape is not
    specified, the default assumes a vector-shaped input, i.e.,
    event_shape_in = list(-1).
  * The Reshape bijector automatically broadcasts over the leftmost
  dimensions of its input (sample_shape and batch_shape); only
  the rightmost event_ndims_in dimensions are reshaped. The
  number of dimensions to reshape is inferred from the provided
  event_shape_in (event_ndims_in = length(event_shape_in)).
}
