% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_scale_matvec_linear_operator}
\alias{tfb_scale_matvec_linear_operator}
\title{Compute \code{Y = g(X; scale) = scale @ X}.}
\usage{
tfb_scale_matvec_linear_operator(scale, adjoint = FALSE,
  validate_args = FALSE, name = "scale_matvec_linear_operator")
}
\arguments{
\item{scale}{Subclass of \code{LinearOperator}. Represents the (batch, non-singular)
linear transformation by which the \code{Bijector} transforms inputs.}

\item{adjoint}{\code{logical} indicating whether to use the \code{scale} matrix as
specified or its adjoint. Default value: \code{FALSE}.}

\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
a bijector instance.
}
\description{
\code{scale} is a \code{LinearOperator}.
If \code{X} is a scalar then the forward transformation is: \code{scale * X}
where \code{*} denotes broadcasted elementwise product.
}
