% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_scale_matvec_lu}
\alias{tfb_scale_matvec_lu}
\title{Matrix-vector multiply using LU decomposition.}
\usage{
tfb_scale_matvec_lu(lower_upper, permutation, validate_args = FALSE,
  name = NULL)
}
\arguments{
\item{lower_upper}{The LU factorization as returned by \code{tf$linalg$lu}.}

\item{permutation}{The LU factorization permutation as returned by \code{tf$linalg$lu}.}

\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
a bijector instance.
}
\description{
This bijector is identical to the "Convolution1x1" used in Glow (Kingma and Dhariwal, 2018).
}
\section{References}{

\itemize{
\item \href{https://arxiv.org/abs/1807.03039}{Diederik P. Kingma, Prafulla Dhariwal. Glow: Generative Flow with Invertible 1x1 Convolutions. arXiv preprint arXiv:1807.03039, 2018.}
}
}

