% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_shift}
\alias{tfb_shift}
\title{Compute \code{Y = g(X; shift) = X + shift}.}
\usage{
tfb_shift(shift, validate_args = FALSE, name = "shift")
}
\arguments{
\item{shift}{floating-point tensor}

\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
a bijector instance.
}
\description{
where \code{shift} is a numeric \code{Tensor}.
}
