% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_softsign}
\alias{tfb_softsign}
\title{Bijector which computes Y = g(X) = X / (1 + |X|).}
\usage{
tfb_softsign(validate_args = FALSE, name = "softsign")
}
\arguments{
\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The softsign Bijector has the following two useful properties:
* The domain is all real numbers
* softsign(x) approx sgn(x), for large |x|.
}
