% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_bernoulli}
\alias{tfd_bernoulli}
\title{The Bernoulli distribution class.}
\usage{
tfd_bernoulli(logits = NULL, probs = NULL, dtype = tf$int32,
  validate_args = FALSE, allow_nan_stats = TRUE, name = "Bernoulli")
}
\arguments{
\item{logits}{An N-D Tensor representing the log-odds of a 1 event. Each entry in the Tensor
parametrizes an independent Bernoulli distribution where the probability of an event
is sigmoid(logits). Only one of logits or probs should be passed in.}

\item{probs}{An N-D Tensor representing the probability of a 1 event. Each entry in the Tensor
parameterizes an independent Bernoulli distribution. Only one of logits or probs
should be passed in.}

\item{dtype}{The type of the event samples. Default: int32.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The Bernoulli distribution class.
}
\seealso{
Other distributions: \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
