% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_blockwise}
\alias{tfd_blockwise}
\title{Blockwise distribution}
\usage{
tfd_blockwise(distributions, dtype_override = NULL,
  validate_args = FALSE, allow_nan_stats = FALSE, name = "Blockwise")
}
\arguments{
\item{distributions}{list of Distribution instances. All distribution instances
must have the same batch_shape and all must have `event_ndims==1``, i.e., be
vector-variate distributions.}

\item{dtype_override}{samples of distributions will be cast to this dtype. If
unspecified, all distributions must have the same dtype. Default value:
\code{NULL} (i.e., do not cast).}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, statistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
a distribution instance.
}
\description{
Blockwise distribution
}
