% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{tfd_entropy}
\alias{tfd_entropy}
\title{Shannon entropy in nats.}
\usage{
tfd_entropy(distribution, name = "entropy")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a Tensor of shape sample_shape(x) + self$batch_shape with values of type self$dtype.
}
\description{
Shannon entropy in nats.
}
