% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{tfd_log_prob}
\alias{tfd_log_prob}
\title{Log probability density/mass function.}
\usage{
tfd_log_prob(distribution, value, name = "log_prob")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{value}{`float` or `double` `Tensor`.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Log probability density/mass function.
}
