% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_mixture}
\alias{tfd_mixture}
\title{Mixture distribution.}
\usage{
tfd_mixture(cat, components, validate_args = FALSE,
  allow_nan_stats = TRUE, use_static_graph = FALSE, name = "Mixture")
}
\arguments{
\item{cat}{A \code{Categorical} distribution instance, representing the probabilities
of \code{distributions}.}

\item{components}{A list or tuple of \code{Distribution} instances.
Each instance must have the same type, be defined on the same domain,
and have matching \code{event_shape} and \code{batch_shape}.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{use_static_graph}{Calls to \code{sample} will not rely on dynamic tensor
indexing, allowing for some static graph compilation optimizations, but
at the expense of sampling all underlying distributions in the mixture.
(Possibly useful when running on TPUs). Default value: \code{FALSE} (i.e., use dynamic indexing).}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The \code{Mixture} object implements batched mixture distributions.
The mixture model is defined by a \code{Categorical} distribution (the mixture)
and a list of \code{Distribution} objects.
}
\details{
Methods supported include \code{tfd_log_prob}, \code{tfd_prob}, \code{tfd_mean}, \code{tfd_sample},
and \code{entropy_lower_bound}.
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_categorical}},
  \code{\link{tfd_exponential}}, \code{\link{tfd_gamma}},
  \code{\link{tfd_horseshoe}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_inverse_gamma}},
  \code{\link{tfd_inverse_gaussian}},
  \code{\link{tfd_joint_distribution_sequential}},
  \code{\link{tfd_kumaraswamy}}, \code{\link{tfd_laplace}},
  \code{\link{tfd_linear_gaussian_state_space_model}},
  \code{\link{tfd_lkj}}, \code{\link{tfd_log_normal}},
  \code{\link{tfd_logistic}},
  \code{\link{tfd_mixture_same_family}},
  \code{\link{tfd_multinomial}},
  \code{\link{tfd_multivariate_normal_diag_plus_low_rank}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_multivariate_normal_full_covariance}},
  \code{\link{tfd_multivariate_normal_linear_operator}},
  \code{\link{tfd_multivariate_normal_tri_l}},
  \code{\link{tfd_multivariate_student_t_linear_operator}},
  \code{\link{tfd_negative_binomial}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_pareto}},
  \code{\link{tfd_poisson_log_normal_quadrature_compound}},
  \code{\link{tfd_poisson}}, \code{\link{tfd_quantized}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_sinh_arcsinh}},
  \code{\link{tfd_student_t_process}},
  \code{\link{tfd_student_t}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_triangular}},
  \code{\link{tfd_truncated_normal}},
  \code{\link{tfd_uniform}},
  \code{\link{tfd_variational_gaussian_process}},
  \code{\link{tfd_vector_diffeomixture}},
  \code{\link{tfd_vector_exponential_diag}},
  \code{\link{tfd_vector_exponential_linear_operator}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
