% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{tfd_mode}
\alias{tfd_mode}
\title{Mean.}
\usage{
tfd_mode(distribution, name = "mode")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a Tensor of shape sample_shape(x) + self$batch_shape with values of type self$dtype.
}
\description{
Mean.
}
