% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_normal}
\alias{tfd_normal}
\title{The Normal distribution with loc and scale parameters}
\usage{
tfd_normal(loc, scale, validate_args = FALSE, allow_nan_stats = TRUE,
  name = "Normal")
}
\arguments{
\item{loc}{Floating point tensor; the means of the distribution(s).}

\item{scale}{loating point tensor; the stddevs of the distribution(s).
Must contain only positive values.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
A normal distribution
}
\description{
The Normal distribution with loc and scale parameters
}
\seealso{
Other distributions: \code{\link{tfd_independent}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
