% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_poisson}
\alias{tfd_poisson}
\title{Poisson distribution.}
\usage{
tfd_poisson(rate = NULL, log_rate = NULL,
  interpolate_nondiscrete = TRUE, validate_args = FALSE,
  allow_nan_stats = TRUE, name = "Poisson")
}
\arguments{
\item{rate}{Floating point tensor, the rate parameter. \code{rate} must be positive.
Must specify exactly one of \code{rate} and \code{log_rate}.}

\item{log_rate}{Floating point tensor, the log of the rate parameter.
Must specify exactly one of \code{rate} and \code{log_rate}.}

\item{interpolate_nondiscrete}{Logical. When \code{FALSE},
\code{log_prob} returns \code{-inf} (and \code{prob} returns \code{0}) for non-integer
inputs. When \code{TRUE}, \code{log_prob} evaluates the continuous function
\code{k * log_rate - lgamma(k+1) - rate}, which matches the Poisson pmf
at integer arguments \code{k} (note that this function is not itself
a normalized probability log-density). Default value: \code{TRUE}.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The Poisson distribution is parameterized by an event \code{rate} parameter.
}
\details{
Mathematical Details

The probability mass function (pmf) is,\preformatted{pmf(k; lambda, k >= 0) = (lambda^k / k!) / Z
Z = exp(lambda).
}

where \code{rate = lambda} and \code{Z} is the normalizing constant.
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_negative_binomial}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_pareto}},
  \code{\link{tfd_poisson_log_normal_quadrature_compound}},
  \code{\link{tfd_quantized}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_sinh_arcsinh}},
  \code{\link{tfd_student_t_process}},
  \code{\link{tfd_student_t}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_triangular}},
  \code{\link{tfd_truncated_normal}},
  \code{\link{tfd_uniform}},
  \code{\link{tfd_variational_gaussian_process}},
  \code{\link{tfd_vector_diffeomixture}},
  \code{\link{tfd_vector_exponential_diag}},
  \code{\link{tfd_vector_exponential_linear_operator}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
