% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_relaxed_bernoulli}
\alias{tfd_relaxed_bernoulli}
\title{RelaxedBernoulli distribution with temperature and logits parameters.}
\usage{
tfd_relaxed_bernoulli(temperature, logits = NULL, probs = NULL,
  validate_args = FALSE, allow_nan_stats = TRUE,
  name = "RelaxedBernoulli")
}
\arguments{
\item{temperature}{An 0-D Tensor, representing the temperature of a set of RelaxedBernoulli distributions.
The temperature should be positive.}

\item{logits}{An N-D Tensor representing the log-odds of a positive event. Each entry in the Tensor
parametrizes an independent RelaxedBernoulli distribution where the probability of an event is sigmoid(logits).
Only one of logits or probs should be passed in.}

\item{probs}{AAn N-D Tensor representing the probability of a positive event. Each entry in the Tensor parameterizes
an independent Bernoulli distribution. Only one of logits or probs should be passed in.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The RelaxedBernoulli is a distribution over the unit interval (0,1), which continuously approximates a Bernoulli.
The degree of approximation is controlled by a temperature: as the temperature goes to 0 the RelaxedBernoulli
becomes discrete with a distribution described by the logits or probs parameters, as the temperature goes to
infinity the RelaxedBernoulli becomes the constant distribution that is identically 0.5.
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
