% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{tfd_sample}
\alias{tfd_sample}
\title{Generate samples of the specified shape.}
\usage{
tfd_sample(distribution, sample_shape = list(), seed = NULL,
  name = "sample")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{sample_shape}{0D or 1D `int32` `Tensor`. Shape of the generated samples.}

\item{seed}{integer seed for RNG}

\item{name}{name to give to the op.}
}
\value{
a `Tensor` with prepended dimensions `sample_shape`.
}
\description{
Note that a call to `tfd_sample()` without arguments will generate a single sample.
}
