% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{tfd_stddev}
\alias{tfd_stddev}
\title{Standard deviation.}
\usage{
tfd_stddev(distribution, name = "stddev")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a Tensor of shape sample_shape(x) + self$batch_shape with values of type self$dtype.
}
\description{
Standard deviation is defined as, stddev = E[(X - E[X])**2]**0.5
#' where X is the random variable associated with this distribution, E denotes expectation,
and Var$shape = batch_shape + event_shape.
}
