% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_student_t}
\alias{tfd_student_t}
\title{Student's t-distribution.}
\usage{
tfd_student_t(df, loc, scale, validate_args = FALSE,
  allow_nan_stats = TRUE, name = "StudentT")
}
\arguments{
\item{df}{Floating-point \code{Tensor}. The degrees of freedom of the
distribution(s). \code{df} must contain only positive values.}

\item{loc}{Floating-point \code{Tensor}. The mean(s) of the distribution(s).}

\item{scale}{Floating-point \code{Tensor}. The scaling factor(s) for the
distribution(s). Note that \code{scale} is not technically the standard
deviation of this distribution but has semantics more similar to
standard deviation than variance.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
This distribution has parameters: degree of freedom \code{df}, location \code{loc}, and \code{scale}.
Notice that \code{scale} has semantics more similar to standard deviation than
variance. However it is not actually the std. deviation; the Student's
t-distribution std. dev. is \code{scale sqrt(df / (df - 2))} when \code{df > 2}.
}
\details{
Samples of this distribution are reparameterized (pathwise differentiable).
The derivatives are computed using the approach described in the paper
\href{https://arxiv.org/abs/1805.08498}{Michael Figurnov, Shakir Mohamed, Andriy Mnih. Implicit Reparameterization Gradients, 2018}
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_triangular}},
  \code{\link{tfd_truncated_normal}},
  \code{\link{tfd_uniform}},
  \code{\link{tfd_variational_gaussian_process}},
  \code{\link{tfd_vector_diffeomixture}},
  \code{\link{tfd_vector_exponential_diag}},
  \code{\link{tfd_vector_exponential_linear_operator}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
