% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{tfd_survival_function}
\alias{tfd_survival_function}
\title{Survival function.}
\usage{
tfd_survival_function(distribution, value, name = "survival_function")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{value}{`float` or `double` `Tensor`.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Given random variable `X`, the survival function is defined:
`survival_function(x) = P[X > x] = 1 - P[X <= x] = 1 - cdf(x).``
}
