% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_transformed}
\alias{tfd_transformed}
\title{A Transformed Distribution.}
\usage{
tfd_transformed(distribution, bijector, batch_shape = NULL,
  event_shape = NULL, validate_args = FALSE, name = NULL)
}
\arguments{
\item{distribution}{The base distribution instance to transform. Typically an instance of Distribution.}

\item{bijector}{The object responsible for calculating the transformation. Typically an instance of Bijector.}

\item{batch_shape}{integer vector Tensor which overrides distribution batch_shape;
valid only if distribution.is_scalar_batch().}

\item{event_shape}{integer vector Tensor which overrides distribution event_shape;
valid only if distribution.is_scalar_event().}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{name}{The name for ops managed by the distribution.  Default value: bijector.name + distribution.name.}
}
\description{
A TransformedDistribution models \code{p(y)} given a base distribution \code{p(x)},
and a deterministic, invertible, differentiable transform,\code{Y = g(X)}. The
transform is typically an instance of the Bijector class and the base
distribution is typically an instance of the Distribution class.
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_variational_gaussian_process}},
  \code{\link{tfd_vector_diffeomixture}},
  \code{\link{tfd_vector_exponential_diag}},
  \code{\link{tfd_vector_exponential_linear_operator}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
