% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_uniform}
\alias{tfd_uniform}
\title{Uniform distribution with \code{low} and \code{high} parameters.}
\usage{
tfd_uniform(low = 0, high = 1, validate_args = FALSE,
  allow_nan_stats = TRUE, name = "Uniform")
}
\arguments{
\item{low}{Floating point tensor, lower boundary of the output interval. Must
have \code{low < high}.}

\item{high}{Floating point tensor, upper boundary of the output interval. Must
have \code{low < high}.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The parameters \code{low} and \code{high} must be shaped in a way that supports
broadcasting (e.g., \code{high - low} is a valid operation).
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_student_t}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_triangular}},
  \code{\link{tfd_truncated_normal}},
  \code{\link{tfd_variational_gaussian_process}},
  \code{\link{tfd_vector_diffeomixture}},
  \code{\link{tfd_vector_exponential_diag}},
  \code{\link{tfd_vector_exponential_linear_operator}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
