% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{tfd_variance}
\alias{tfd_variance}
\title{Variance.}
\usage{
tfd_variance(distribution, name = "variance")
}
\arguments{
\item{distribution}{The distribution being used.}

\item{name}{String prepended to names of ops created by this function.}
}
\value{
a Tensor of shape sample_shape(x) + self$batch_shape with values of type self$dtype.
}
\description{
Variance is defined as, Var = E[(X - E[X])**2]
where X is the random variable associated with this distribution, E denotes expectation,
and Var$shape = batch_shape + event_shape.
}
