% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_vector_laplace_diag}
\alias{tfd_vector_laplace_diag}
\title{The vectorization of the Laplace distribution on R^k.}
\usage{
tfd_vector_laplace_diag(loc = NULL, scale_diag = NULL,
  scale_identity_multiplier = NULL, validate_args = FALSE,
  allow_nan_stats = TRUE, name = "VectorLaplaceDiag")
}
\arguments{
\item{loc}{Floating-point Tensor. If this is set to NULL, loc is
implicitly 0. When specified, may have shape [B1, ..., Bb, k] where
b >= 0 and k is the event size.}

\item{scale_diag}{Non-zero, floating-point Tensor representing a diagonal
matrix added to scale. May have shape [B1, ..., Bb, k], b >= 0,
and characterizes b-batches of k x k diagonal matrices added to
scale. When both scale_identity_multiplier and scale_diag are
NULL then scale is the Identity.}

\item{scale_identity_multiplier}{Non-zero, floating-point Tensor representing
a scaled-identity-matrix added to scale. May have shape
[B1, ..., Bb], b >= 0, and characterizes b-batches of scaled
k x k identity matrices added to scale. When both
scale_identity_multiplier and scale_diag are NULL then scale is
the Identity.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The vector laplace distribution is defined over R^k, and parameterized by
a (batch of) length-k loc vector (the means) and a (batch of) k x k
scale matrix:  covariance = 2 * scale @ scale.T, where @ denotes
matrix-multiplication.
}
\details{
About VectorLaplace and Vector distributions in TensorFlow.
The VectorLaplace is a non-standard distribution that has useful properties.
The marginals Y_1, ..., Y_k are *not* Laplace random variables, due to
the fact that the sum of Laplace random variables is not Laplace.
Instead, Y is a vector whose components are linear combinations of Laplace
random variables.  Thus, Y lives in the vector space generated by vectors
of Laplace distributions.  This allows the user to decide the mean and
covariance (by setting loc and scale), while preserving some properties of
the Laplace distribution.  In particular, the tails of Y_i will be (up to
polynomial factors) exponentially decaying.
To see this last statement, note that the pdf of Y_i is the convolution of
the pdf of k independent Laplace random variables.  One can then show by
induction that distributions with exponential (up to polynomial factors) tails
are closed under convolution.
}
\seealso{
Other distributions: \code{\link{tfd_independent}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
