% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_vector_laplace_linear_operator}
\alias{tfd_vector_laplace_linear_operator}
\title{The vectorization of the Laplace distribution on \code{R^k}.}
\usage{
tfd_vector_laplace_linear_operator(loc = NULL, scale = NULL,
  validate_args = FALSE, allow_nan_stats = TRUE,
  name = "VectorLaplaceLinearOperator")
}
\arguments{
\item{loc}{Floating-point Tensor. If this is set to NULL, loc is
implicitly 0. When specified, may have shape \code{[B1, ..., Bb, k]} where
b >= 0 and k is the event size.}

\item{scale}{Instance of LinearOperator with same dtype as loc and shape
\code{[B1, ..., Bb, k, k]}.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The vector laplace distribution is defined over \code{R^k}, and parameterized by
a (batch of) length-k loc vector (the means) and a (batch of) k x k
scale matrix:  \code{covariance = 2 * scale @ scale.T}, where @ denotes
matrix-multiplication.
}
\details{
About VectorLaplace and Vector distributions in TensorFlow.
The VectorLaplace is a non-standard distribution that has useful properties.
The marginals Y_1, ..., Y_k are \emph{not} Laplace random variables, due to
the fact that the sum of Laplace random variables is not Laplace.
Instead, Y is a vector whose components are linear combinations of Laplace
random variables.  Thus, Y lives in the vector space generated by vectors
of Laplace distributions.  This allows the user to decide the mean and
covariance (by setting loc and scale), while preserving some properties of
the Laplace distribution.  In particular, the tails of Y_i will be (up to
polynomial factors) exponentially decaying.
To see this last statement, note that the pdf of Y_i is the convolution of
the pdf of k independent Laplace random variables.  One can then show by
induction that distributions with exponential (up to polynomial factors) tails
are closed under convolution.

The batch_shape is the broadcast shape between loc and scale
arguments.
The event_shape is given by last dimension of the matrix implied by
scale. The last dimension of loc (if provided) must broadcast with this.
Recall that covariance = 2 * scale @ scale.T.
Additional leading dimensions (if any) will index batches.
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
