% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_vector_sinh_arcsinh_diag}
\alias{tfd_vector_sinh_arcsinh_diag}
\title{The (diagonal) SinhArcsinh transformation of a distribution on \code{R^k}.}
\usage{
tfd_vector_sinh_arcsinh_diag(loc = NULL, scale_diag = NULL,
  scale_identity_multiplier = NULL, skewness = NULL,
  tailweight = NULL, distribution = NULL, validate_args = FALSE,
  allow_nan_stats = TRUE, name = "VectorSinhArcsinhDiag")
}
\arguments{
\item{loc}{Floating-point Tensor. If this is set to NULL, loc is
implicitly 0. When specified, may have shape \code{[B1, ..., Bb, k]} where
b >= 0 and k is the event size.}

\item{scale_diag}{Non-zero, floating-point Tensor representing a diagonal
matrix added to scale. May have shape \code{[B1, ..., Bb, k]}, b >= 0,
and characterizes b-batches of k x k diagonal matrices added to
scale. When both scale_identity_multiplier and scale_diag are
NULL then scale is the Identity.}

\item{scale_identity_multiplier}{Non-zero, floating-point Tensor representing
a scale-identity-matrix added to scale. May have shape
\code{[B1, ..., Bb]}, b >= 0, and characterizes b-batches of scale
k x k identity matrices added to scale. When both
scale_identity_multiplier and scale_diag are NULL then scale
is the Identity.}

\item{skewness}{Skewness parameter.  floating-point Tensor with shape
broadcastable with event_shape.}

\item{tailweight}{Tailweight parameter.  floating-point Tensor with shape
broadcastable with event_shape.}

\item{distribution}{\code{tf$distributions$Distribution}-like instance. Distribution from which k
iid samples are used as input to transformation F.  Default is
tfd_normal(loc = 0, scale = 1).
Must be a scalar-batch, scalar-event distribution.  Typically
distribution$reparameterization_type = FULLY_REPARAMETERIZED or it is
a function of non-trainable parameters. WARNING: If you backprop through
a VectorSinhArcsinhDiag sample and distribution is not
FULLY_REPARAMETERIZED yet is a function of trainable variables, then
the gradient will be incorrect!}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
This distribution models a random vector \code{Y = (Y1,...,Yk)}, making use of
a SinhArcsinh transformation (which has adjustable tailweight and skew),
a rescaling, and a shift.
The SinhArcsinh transformation of the Normal is described in great depth in
\href{https://www.jstor.org/stable/27798865}{Sinh-arcsinh distributions}.
Here we use a slightly different parameterization, in terms of tailweight
and skewness.  Additionally we allow for distributions other than Normal,
and control over scale as well as a "shift" parameter loc.
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_vector_exponential_diag}},
  \code{\link{tfd_vector_exponential_linear_operator}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
