% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_von_mises}
\alias{tfd_von_mises}
\title{The von Mises distribution over angles.}
\usage{
tfd_von_mises(loc, concentration, validate_args = FALSE,
  allow_nan_stats = TRUE, name = "VonMises")
}
\arguments{
\item{loc}{Floating point tensor, the circular means of the distribution(s).}

\item{concentration}{Floating point tensor, the level of concentration of the
distribution(s) around loc. Must take non-negative values.
concentration = 0 defines a Uniform distribution, while
concentration = +inf indicates a Deterministic distribution at loc.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The von Mises distribution is a univariate directional distribution.
Similarly to Normal distribution, it is a maximum entropy distribution.
The samples of this distribution are angles, measured in radians.
They are 2 pi-periodic: x = 0 and x = 2pi are equivalent.
This means that the density is also 2 pi-periodic.
The generated samples, however, are guaranteed to be in \code{[-pi, pi)} range.
When concentration = 0, this distribution becomes a Uniform distribuion on
the \code{[-pi, pi)} domain.
}
\details{
The von Mises distribution is a special case of von Mises-Fisher distribution
for n=2. However, the TFP's VonMisesFisher implementation represents the
samples and location as (x, y) points on a circle, while VonMises represents
them as scalar angles.

Mathematical details
The probability density function (pdf) of this distribution is,\preformatted{pdf(x; loc, concentration) = exp(concentration cos(x - loc)) / Z
Z = 2 * pi * I_0 (concentration)
}

where:
\itemize{
\item \code{I_0 (concentration)} is the modified Bessel function of order zero;
\item \code{loc} the circular mean of the distribution, a scalar. It can take arbitrary
values, but it is 2pi-periodic: loc and loc + 2pi result in the same
distribution.
\item \code{concentration >= 0} parameter is the concentration parameter. When
\code{concentration = 0},
this distribution becomes a Uniform distribution on [-pi, pi).
}

The parameters loc and concentration must be shaped in a way that
supports broadcasting (e.g. loc + concentration is a valid operation).
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_poisson_log_normal_quadrature_compound}},
  \code{\link{tfd_poisson}}, \code{\link{tfd_quantized}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_sinh_arcsinh}},
  \code{\link{tfd_student_t_process}},
  \code{\link{tfd_student_t}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_triangular}},
  \code{\link{tfd_truncated_normal}},
  \code{\link{tfd_uniform}},
  \code{\link{tfd_variational_gaussian_process}},
  \code{\link{tfd_vector_diffeomixture}},
  \code{\link{tfd_vector_exponential_diag}},
  \code{\link{tfd_vector_exponential_linear_operator}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises_fisher}},
  \code{\link{tfd_wishart}}, \code{\link{tfd_zipf}}
}
\concept{distributions}
