% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{tfd_von_mises_fisher}
\alias{tfd_von_mises_fisher}
\title{The von Mises-Fisher distribution over unit vectors on \code{S^{n-1}}.}
\usage{
tfd_von_mises_fisher(df, mean_direction, concentration,
  validate_args = FALSE, allow_nan_stats = TRUE,
  name = "VonMisesFisher")
}
\arguments{
\item{mean_direction}{Floating-point Tensor with shape \code{[B1, ... Bn, D]}.
A unit vector indicating the mode of the distribution, or the
unit-normalized direction of the mean. (This is \emph{not} in general the
mean of the distribution; the mean is not generally in the support of
the distribution.) NOTE: D is currently restricted to <= 5.}

\item{concentration}{Floating-point Tensor having batch shape \code{[B1, ... Bn]}
broadcastable with mean_direction. The level of concentration of
samples around the mean_direction. concentration=0 indicates a
uniform distribution over the unit hypersphere, and concentration=+inf
indicates a Deterministic distribution (delta function) at mean_direction.}

\item{validate_args}{Logical, default FALSE. When TRUE distribution parameters are checked
for validity despite possibly degrading runtime performance. When FALSE invalid inputs may
silently render incorrect outputs. Default value: FALSE.}

\item{allow_nan_stats}{Logical, default TRUE. When TRUE, tatistics (e.g., mean, mode, variance)
use the value NaN to indicate the result is undefined. When FALSE, an exception is raised if
one or more of the statistic's batch members are undefined.}

\item{name}{name prefixed to Ops created by this class.}
}
\description{
The von Mises-Fisher distribution is a directional distribution over vectors
on the unit hypersphere \code{S^{n-1}} embedded in n dimensions \code{(R^n)}.
}
\details{
NOTE: Currently only n in {2, 3, 4, 5} are supported. For n=5 some numerical
instability can occur for low concentrations (<.01).
}
\seealso{
Other distributions: \code{\link{tfd_bernoulli}},
  \code{\link{tfd_independent}},
  \code{\link{tfd_multivariate_normal_diag}},
  \code{\link{tfd_normal}},
  \code{\link{tfd_one_hot_categorical}},
  \code{\link{tfd_relaxed_bernoulli}},
  \code{\link{tfd_relaxed_one_hot_categorical}},
  \code{\link{tfd_transformed}},
  \code{\link{tfd_vector_laplace_diag}},
  \code{\link{tfd_vector_laplace_linear_operator}},
  \code{\link{tfd_vector_sinh_arcsinh_diag}},
  \code{\link{tfd_von_mises}}, \code{\link{tfd_wishart}},
  \code{\link{tfd_zipf}}
}
\concept{distributions}
