% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{tfp}
\alias{tfp}
\title{Handle to the \code{tensorflow_probability} module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
tfp
}
\value{
Module(tensorflow_probability)
}
\description{
Handle to the \code{tensorflow_probability} module
}
\keyword{datasets}
