% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-methods.R
\name{variance.tensorflow_probability.python.distributions.distribution.Distribution}
\alias{variance.tensorflow_probability.python.distributions.distribution.Distribution}
\title{Variance.}
\usage{

  \method{variance}{tensorflow_probability.python.distributions.distribution.Distribution}(distribution,
  name = "variance")
}
\arguments{
\item{name}{String prepended to names of ops created by this function.}
}
\value{
a `Tensor` of shape `sample_shape(x) + self$batch_shape` with values of type `self$dtype`.
}
\description{
Variance is defined as, `Var = E[(X - E[X])**2]`
where `X` is the random variable associated with this distribution, `E` denotes expectation,
and `Var$shape = batch_shape + event_shape`.
}
