% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi-functions.R
\name{vi_kl_reverse}
\alias{vi_kl_reverse}
\title{The reverse Kullback-Leibler Csiszar-function in log-space.}
\usage{
vi_kl_reverse(logu, self_normalized = FALSE, name = NULL)
}
\arguments{
\item{logu}{\code{float}-like \code{Tensor} representing \code{log(u)} from above.}

\item{self_normalized}{\code{logical} indicating whether \code{f'(u=1)=0}. When
\code{f'(u=1)=0} the implied Csiszar f-Divergence remains non-negative even
when \code{p, q} are unnormalized measures.}

\item{name}{name prefixed to Ops created by this function.}
}
\description{
A Csiszar-function is a member of \code{F = { f:R_+ to R : f convex }}.
}
\details{
When \code{self_normalized = TRUE}, the KL-reverse Csiszar-function is \code{f(u) = -log(u) + (u - 1)}.
When \code{self_normalized = False} the \code{(u - 1)} term is omitted.
Observe that as an f-Divergence, this Csiszar-function implies: \code{D_f[p, q] = KL[q, p]}

The KL is "reverse" because in maximum likelihood we think of minimizing \code{q} as in \code{KL[p, q]}.

Warning: when self_normalized = True\code{this function makes non-log-space calculations and may therefore be numerically unstable for}|logu| >> 0`.
}
\seealso{
Other vi-functions: \code{\link{vi_amari_alpha}}
}
\concept{vi-functions}
