% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi-functions.R
\name{vi_log1p_abs}
\alias{vi_log1p_abs}
\title{The log1p-abs Csiszar-function in log-space}
\usage{
vi_log1p_abs(logu, name = NULL)
}
\arguments{
\item{logu}{\code{float}-like \code{Tensor} representing \code{log(u)} from above.}

\item{name}{name prefixed to Ops created by this function.}
}
\value{
log1p_abs_of_u: \code{float}-like \code{Tensor} of the Csiszar-function
evaluated at \code{u = exp(logu)}.
}
\description{
A Csiszar-function is a member of \verb{F = \{ f:R_+ to R : f convex \}}.
}
\details{
The Log1p-Abs Csiszar-function is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{f(u) = u**(sign(u-1)) - 1
}\if{html}{\out{</div>}}

This function is so-named because it was invented from the following recipe.
Choose a convex function g such that g(0)=0 and solve for f:

\if{html}{\out{<div class="sourceCode">}}\preformatted{log(1 + f(u)) = g(log(u)).
<=>
f(u) = exp(g(log(u))) - 1
}\if{html}{\out{</div>}}

That is, the graph is identically \code{g} when y-axis is \code{log1p}-domain and x-axis
is \code{log}-domain.

Warning: this function makes non-log-space calculations and may
therefore be numerically unstable for \verb{|logu| >> 0}.
}
\seealso{
Other vi-functions: 
\code{\link{vi_amari_alpha}()},
\code{\link{vi_arithmetic_geometric}()},
\code{\link{vi_chi_square}()},
\code{\link{vi_csiszar_vimco}()},
\code{\link{vi_dual_csiszar_function}()},
\code{\link{vi_fit_surrogate_posterior}()},
\code{\link{vi_jeffreys}()},
\code{\link{vi_jensen_shannon}()},
\code{\link{vi_kl_forward}()},
\code{\link{vi_kl_reverse}()},
\code{\link{vi_modified_gan}()},
\code{\link{vi_monte_carlo_variational_loss}()},
\code{\link{vi_pearson}()},
\code{\link{vi_squared_hellinger}()},
\code{\link{vi_symmetrized_csiszar_function}()}
}
\concept{vi-functions}
