% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi-functions.R
\name{vi_t_power}
\alias{vi_t_power}
\title{The T-Power Csiszar-function in log-space}
\usage{
vi_t_power(logu, t, self_normalized = FALSE, name = NULL)
}
\arguments{
\item{logu}{\code{float}-like \code{Tensor} representing \code{log(u)} from above.}

\item{t}{\code{Tensor} of same \code{dtype} as \code{logu} and broadcastable shape.}

\item{self_normalized}{\code{logical} indicating whether \code{f'(u=1)=0}. When
\code{f'(u=1)=0} the implied Csiszar f-Divergence remains non-negative even
when \code{p, q} are unnormalized measures.}

\item{name}{name prefixed to Ops created by this function.}
}
\value{
t_power_of_u: \code{float}-like \code{Tensor} of the Csiszar-function
evaluated at \code{u = exp(logu)}.#'
}
\description{
A Csiszar-function is a member of \code{F = { f:R_+ to R : f convex }}.
}
\seealso{
Other vi-functions#': \code{\link{vi_total_variation}},
  \code{\link{vi_triangular}}
}
\concept{vi-functions#'}
