% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi-functions.R
\name{vi_total_variation}
\alias{vi_total_variation}
\title{The Total Variation Csiszar-function in log-space}
\usage{
vi_total_variation(logu, name = NULL)
}
\arguments{
\item{logu}{\code{float}-like \code{Tensor} representing \code{log(u)} from above.}

\item{name}{name prefixed to Ops created by this function.}
}
\value{
total_variation_of_u: \code{float}-like \code{Tensor} of the Csiszar-function
evaluated at \code{u = exp(logu)}.
}
\description{
A Csiszar-function is a member of \verb{F = \{ f:R_+ to R : f convex \}}.
}
\details{
The Total-Variation Csiszar-function is:\preformatted{f(u) = 0.5 |u - 1|
}

Warning: this function makes non-log-space calculations and may therefore be
numerically unstable for \verb{|logu| >> 0}.
}
\seealso{
Other vi-functions#': 
\code{\link{vi_t_power}()},
\code{\link{vi_triangular}()}
}
\concept{vi-functions#'}
