% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_fit}
\alias{glm_fit}
\title{Runs multiple Fisher scoring steps}
\usage{
glm_fit(x, ...)
}
\arguments{
\item{x}{float-like, matrix-shaped Tensor where each row represents a sample's
features.}

\item{...}{other arguments passed to specific methods.}
}
\value{
A \code{glm_fit} object with parameter estimates, number of iterations,
etc.
}
\description{
Runs multiple Fisher scoring steps
}
\seealso{
\code{\link[=glm_fit.tensorflow.tensor]{glm_fit.tensorflow.tensor()}}
}
