% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tfprobability}
\alias{install_tfprobability}
\title{Installs TensorFlow Probability}
\usage{
install_tfprobability(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  version = "default",
  tensorflow = "default",
  extra_packages = NULL,
  ...
)
}
\arguments{
\item{method}{Installation method ("virtualenv" or "conda")}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{version}{Version of Keras to install. Specify "default" to install
the latest release. Otherwise specify an alternate version (e.g. "2.2.2").}

\item{tensorflow}{TensorFlow version to install. Specify "default" to install
the CPU version of the latest release. Specify "gpu" to install the GPU
version of the latest release.

You can also provide a full major.minor.patch specification (e.g. "1.1.0"),
appending "-gpu" if you want the GPU version (e.g. "1.1.0-gpu").

Alternatively, you can provide the full URL to an installer binary (e.g.
for a nightly binary).}

\item{extra_packages}{Additional PyPI packages to install along with
Keras and TensorFlow.}

\item{...}{Other arguments passed to \code{\link[tensorflow:install_tensorflow]{tensorflow::install_tensorflow()}}.}
}
\value{
invisible
}
\description{
Installs TensorFlow Probability
}
