% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-static-methods.R
\name{params_size_independent_bernoulli}
\alias{params_size_independent_bernoulli}
\title{number of \code{params} needed to create an IndependentBernoulli distribution}
\usage{
params_size_independent_bernoulli(event_size)
}
\arguments{
\item{event_size}{event size of this distribution}
}
\value{
a scalar
}
\description{
number of \code{params} needed to create an IndependentBernoulli distribution
}
