% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bijectors.R
\name{tfb_cumsum}
\alias{tfb_cumsum}
\title{Computes the cumulative sum of a tensor along a specified axis.}
\usage{
tfb_cumsum(axis = -1, validate_args = FALSE, name = "cumsum")
}
\arguments{
\item{axis}{\code{int} indicating the axis along which to compute the cumulative sum.
Note that positive (and zero) values are not supported}

\item{validate_args}{Logical, default FALSE. Whether to validate input with asserts. If validate_args is
FALSE, and the inputs are invalid, correct behavior is not guaranteed.}

\item{name}{name prefixed to Ops created by this class.}
}
\value{
a bijector instance.
}
\description{
Computes the cumulative sum of a tensor along a specified axis.
}
